ezPost/requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.4
django-ckeditor==5.6.1
django-filter==2.0.0
django-js-asset==1.1.0
djangorestframework==3.9.0
idna==2.8
Markdown==3.0.1
oauthlib==2.1.0
Pillow==5.3.0
py2-py3-django-email-as-username==1.7.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
urllib3==1.24.1
